test(proptest): replace tautological tests + wire for_each to parser
- Replace json_fuzzing module (tested serde_json, not Nika) with
  binding_storage_fuzzing that tests ResolvedBindings set/get roundtrip
- Replace test_valid_task_id_passes (regex tautology) with
  test_valid_task_id_accepted_by_parser that wires through parse+analyze
- Wire test_for_each_empty_array through parse()+analyze() instead of
  just serde_yaml::from_str
- Wire test_for_each_non_array through parse()+analyze() instead of
  just serde_yaml::from_str

+++ tools/nika/src/core/backend.rs
@@ -420,9 +420,9 @@ pub struct ChatResponse {
     /// Total duration in nanoseconds.
     pub total_duration: Option<u64>,
     /// Tokens generated.
-    pub eval_count: Option<u32>,
+    pub eval_count: Option<u64>,
     /// Prompt tokens.
-    pub prompt_eval_count: Option<u32>,
+    pub prompt_eval_count: Option<u64>,
 }
 
 impl ChatResponse {
